assistant/requirements.txt
twitchio
pytchat
PyYAML
python-telegram-bot>=20.0
flask
websockets>=12.0
edge-tts
requests
sounddevice
numpy
faster-whisper
opencv-python
easyocr
pillow 